Bugfix game's working directory not set in boot.sh
When games are launched via the boot.sh file, the working directory
wasn't set to the game's root folder which resulted in files not
being found, e.g. game assets. These depend on the working directory
set before the game is started.

Because the game is booted from boot.sh, the fix should be located
there.