os: Re-work installation process, all installation steps in chroot
Seems like this is not a problem (anymore?). Execute all installation
steps including kernel update, gpu driver installation and user
space application installation in the chroot environment.

Therefore, the whole "install on first boot" step is not required
as a separte step allowing you to complete the entire OS setup
on the host/development machine.