Bugfix: Install grub to mounted disk using grub-install
Apparently, the install process of the grub packages changed. It
doesn't include selecting the boot device to install the boot loader
using a UI anymore. On the upside, this allows us to finally create
a fully automated installation without any manual interactions.
The downside, debugging, confirming this was actually removed
and coming up with a fix.